pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neurobalance"
version = "0.1.0"
description = "NEUROBALANCE: Domain-specific neuron activation framework for multimodal text generation"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

authors = [
  { name = "Autin Olom Ogar", email = "austinolomogar@gmail.com" }
]

dependencies = [
  "torch>=2.0.0",
  "transformers>=4.40.0",
  "datasets>=2.18.0",
  "accelerate>=0.29.0",
  "peft>=0.11.0",
  "bitsandbytes>=0.43.0",
  "opencv-python>=4.8.0",
  "matplotlib>=3.7.0",
  "tqdm>=4.66.0",
  "pyyaml>=6.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
  "black>=24.0.0"
]

[tool.setuptools]
packages = ["neurobalance"]

[tool.setuptools.package-data]
neurobalance = ["**/*.yaml", "**/*.json"]
